requirements.txt
albumentations>=1.4.0
numpy>=1.26.0
Pillow>=10.0.0
PyYAML>=6.0.0
scikit-learn>=1.4.0
segmentation-models-pytorch>=0.3.4
torch>=2.2.0
torchvision>=0.17.0
tensorboard>=2.16.0
tqdm>=4.66.0
